+++ app/containers/profilePanel.js
@@ -2,7 +2,16 @@ import React from 'react'
 import { connect } from 'react-redux'
 
 import {
-  hideProfilePanel
+  getUsers,
+  hideProfilePanel,
+  moderationAddAdmin,
+  moderationAddMod,
+  moderationBlock,
+  moderationHide,
+  moderationRemoveAdmin,
+  moderationRemoveMod,
+  moderationUnblock,
+  moderationUnhide
 } from '../actions'
 import Avatar from './avatar'
 
@@ -12,14 +21,75 @@ const mapStateToProps = state => ({
 })
 
 const mapDispatchToProps = dispatch => ({
-  hideProfilePanel: ({ addr }) => dispatch(hideProfilePanel({ addr }))
+  getUsers: ({ addr }) => dispatch(getUsers({ addr })),
+  hideProfilePanel: ({ addr }) => dispatch(hideProfilePanel({ addr })),
+  moderationAddAdmin: ({ addr, channel, reason, userKey }) => dispatch(moderationAddAdmin({ addr, channel, reason, userKey })),
+  moderationAddMod: ({ addr, channel, reason, userKey }) => dispatch(moderationAddMod({ addr, channel, reason, userKey })),
+  moderationBlock: ({ addr, channel, reason, userKey }) => dispatch(moderationBlock({ addr, channel, reason, userKey })),
+  moderationHide: ({ addr, channel, reason, userKey }) => dispatch(moderationHide({ addr, channel, reason, userKey })),
+  moderationRemoveAdmin: ({ addr, channel, reason, userKey }) => dispatch(moderationRemoveAdmin({ addr, channel, reason, userKey })),
+  moderationRemoveMod: ({ addr, channel, reason, userKey }) => dispatch(moderationRemoveMod({ addr, channel, reason, userKey })),
+  moderationUnblock: ({ addr, channel, reason, userKey }) => dispatch(moderationUnblock({ addr, channel, reason, userKey })),
+  moderationUnhide: ({ addr, channel, reason, userKey }) => dispatch(moderationUnhide({ addr, channel, reason, userKey }))
 })
 
 function ProfilePanel (props) {
-  const user = props.user
+  const user = props.getUsers({ addr: props.addr })[props.userKey]
 
-  function onClickHideUser () {
-    console.log('hide', user)
+  function onClickHideUserAll () {
+    props.moderationHide({
+      addr: props.addr,
+      userKey: user.key
+    })
+  }
+
+  function onClickUnhideUserAll () {
+    props.moderationUnhide({
+      addr: props.addr,
+      userKey: user.key
+    })
+  }
+
+  function onClickBlockUserAll () {
+    props.moderationBlock({
+      addr: props.addr,
+      userKey: user.key
+    })
+  }
+
+  function onClickUnblockUserAll () {
+    props.moderationUnblock({
+      addr: props.addr,
+      userKey: user.key
+    })
+  }
+
+  function onClickAddModAll () {
+    props.moderationAddMod({
+      addr: props.addr,
+      userKey: user.key
+    })
+  }
+
+  function onClickRemoveModAll () {
+    props.moderationRemoveMod({
+      addr: props.addr,
+      userKey: user.key
+    })
+  }
+
+  function onClickAddAdminAll () {
+    props.moderationAddAdmin({
+      addr: props.addr,
+      userKey: user.key
+    })
+  }
+
+  function onClickRemoveAdminAll () {
+    props.moderationRemoveAdmin({
+      addr: props.addr,
+      userKey: user.key
+    })
   }
 
   return (
@@ -37,28 +107,57 @@ function ProfilePanel (props) {
             {!user.online &&
               <div title='Offline' className='indicator offline' />}
           </div>
-
           <h1 className='name'>{user.name}</h1>
           <h2 className='key' title={user.key}>{user.key}</h2>
+          <div className='sigilContainer'>
+            {user.isAdmin() && <div className='sigil admin'>Admin</div>}
+            {user.isModerator() && <div className='sigil moderator'>Moderator</div>}
+            {user.isHidden() && <div className='sigil hidden'>Hidden</div>}
+          </div>
         </div>
       </div>
       <div className='section__header'>
         Moderation
       </div>
       <div className='panel__content'>
-        <div className='help__text'>The following features are a work in progress and may not be fully functional yet.</div>
-        <br />
-        <br />
-        <button className='button' onClick={onClickHideUser}>Mute this peer</button>
+        {/* <button className='button' onClick={onClickHideUserAll}>Mute this peer</button>
         <div className='help__text'>Muting a peer hides all of their future messages, while preserving their messages prior to the mute.</div>
         <br />
+        <br /> */}
+        {!user.isHidden() &&
+          <>
+            <button className='button' onClick={onClickHideUserAll}>Hide this peer</button>
+            <div className='help__text'>Hiding a peer hides all of their past and future messages in all channels.</div>
+          </>}
+        {user.isHidden() &&
+          <>
+            <button className='button' onClick={onClickUnhideUserAll}>Unhide this peer</button>
+            <div className='help__text'>Hiding a peer hides all of their past and future messages in all channels.</div>
+          </>}
+        {/* <br />
         <br />
-        <button className='button' onClick={onClickHideUser}>Hide this peer</button>
-        <div className='help__text'>Hiding a peer hides all of their past and future messages.</div>
-        <br />
-        <br />
-        <button className='button' onClick={onClickHideUser}>Block this peer</button>
-        <div className='help__text'>Blocking a peer removes all of their messages from your computer. All past and future messages will not be visible, or even known about.</div>
+        <button className='button' onClick={onClickHideUserAll}>Block this peer</button>
+        <div className='help__text'>Blocking a peer removes all of their messages from your computer. All past and future messages will not be visible, or even known about.</div> */}
+        {!user.isModerator() &&
+          <>
+            <button className='button' onClick={onClickAddModAll}>Add moderator</button>
+            <div className='help__text'>Adding another user as a moderator for you will apply their moderation settings to how you see this cabal.</div>
+          </>}
+        {user.isModerator() &&
+          <>
+            <button className='button' onClick={onClickRemoveModAll}>Remove moderator</button>
+            <div className='help__text'>Adding another user as a moderator for you will apply their moderation settings to how you see this cabal.</div>
+          </>}
+        {/* {!user.isAdmin() &&
+          <>
+            <button className='button' onClick={onClickAddAdminAll}>Add admin</button>
+            <div className='help__text'>Adding another user as an admin for you will apply their moderation settings to how you see this cabal.</div>
+          </>}
+        {user.isAdmin() &&
+          <>
+            <button className='button' onClick={onClickRemoveAdminAll}>Remove admin</button>
+            <div className='help__text'>Adding another user as an admin for you will apply their moderation settings to how you see this cabal.</div>
+          </>} */}
       </div>
     </div>
   )