+++ app/selectors.js
@@ -9,8 +9,14 @@ export const currentCabalSelector = createSelector(
 function sortUsers (users = []) {
   if (Array.isArray(users)) {
     return users.sort((a, b) => {
+      if (a.isHidden() && !b.isHidden()) return 1
+      if (b.isHidden() && !a.isHidden()) return -1
       if (a.online && !b.online) return -1
       if (b.online && !a.online) return 1
+      if (a.isAdmin() && !b.isAdmin()) return -1
+      if (b.isAdmin() && !a.isAdmin()) return 1
+      if (a.isModerator() && !b.isModerator()) return -1
+      if (b.isModerator() && !a.isModerator()) return 1
       if (a.name && !b.name) return -1
       if (b.name && !a.name) return 1
       if (a.name && b.name) return a.name.toLowerCase() < b.name.toLowerCase() ? -1 : 1