+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [5.0.6] - 2020-05-29
+### Added
+- Added basic moderation features for hiding other user's activity.
+- Added indicator for unread messages in collapsed channels list.
+- Optimized incoming message handling.
+### Fixed
+- Fixed desktop notifications appearing from channels you have not joined. 
+
 ## [5.0.5] - 2020-05-19
 ### Added
 - Added "starred/favorite" channels list.